requirements.txt
Pillow
numpy
aiohttp
asyncio
aiofiles